Authenticate user credentials

auth asks the API's server for a session ID (SID), which will later be passed to either query_wos or pull_wos. Note, there are limits on how many session IDs you can get in a given period of time (roughly 5 SIDs in a 5 minute time period).

auth(username = Sys.getenv("WOS_USERNAME"),
  password = Sys.getenv("WOS_PASSWORD"))

Arguments

username
Your username. Specify username = NULL if you want to use IP-based authentication.

password
Your password. Specify password = NULL if you want to use IP-based authentication.

Value

A session ID

Examples

# NOT RUN {
# Pass user credentials in manually:
auth("some_username", password = "some_password")

# Use default of looking for username and password in env variables, so you
# don't have to keep specifying them in your code:
Sys.setenv(WOS_USERNAME = "some_username", WOS_PASSWORD = "some_password")
auth()
# }