Pull data from the InCites API

Important note: The throttling limits on the InCites API are not documented anywhere and are difficult to determine from experience. As such, whenever pull_incites receives a throttling error from the server, it uses exponential backoff (with a maximum wait time of 45 minutes) to determine how long to wait before retrying.

pull_incites(uts, key = Sys.getenv("INCITES_KEY"), as_raw = FALSE, ...)

Arguments

uts
A vector of UTs whose InCites data you would like to download. Each UT is a 15-digit identifier for a given publication. You can specify the UT using only these 15 digits or you can append the 15 digits with "WOS:" (e.g., "000346263300011" or "WOS:000346263300011").

key
The developer key that the server will use for authentication.

as_raw
Do you want the data frame that is returned by the API to be returned to you in its raw form? This option can be useful if the API has changed the format of the data that it is serving, in which case specifying as_raw = TRUE may avoid an error that would otherwise occur during pull_incites's data processing step.

...
Arguments passed along to GET.

Value

A data frame where each row corresponds to a different publication. The definitions for the columns in this data frame can be found online at the API's documentation page (see the DocumentLevelMetricsByUT method details for definitions). Note that the column names are all converted to lowercase by pull_incites and the 0/1 flag variables are converted to booleans. Also note that not all publications indexed in WoS are also indexed in InCites, so you may not get data back for some UTs.

Examples

# NOT RUN {
uts <- c(
  "WOS:000346263300011", "WOS:000362312600021", "WOS:000279885800004",
  "WOS:000294667500003", "WOS:000294946900020", "WOS:000412659200006"
)
pull_incites(uts, key = "some_key")

pull_incites(c("000346263300011", "000362312600021"), key = "some_key")
# }